Start on adding tile tray definitions for Board18 use
At the moment the Board18 output groups all tiles by colour and then
orders them by tile ID. This can give extremely large tile trays (I'm
looking at you, 1858, and your sixty different green tiles) and the
ordering of the tiles can be unhelpful, it's better to have all the
plain track together, and all the towns, and the cities and so on.

The current JSON definition of the game doesn't allow any specification
of different trays or ordering within them. This commit adds a tileTrays
property to the game object, this contains an array of objects, each of
which has a name and a list of tile IDs.